LayoutTests/inspector-protocol/model/content-flow-content-nodes.html

<!doctype html>
<html>
<head>
<style>
.content
{
    -webkit-flow-into: flow;
}
.hidden
{
    -webkit-flow-into: none;
}
</style>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script>
function changeFlowContent()
{
    document.getElementById("contentAddedAtTheTop").classList.remove("hidden");
    // Force a layout to make sure the events are always fired in the expected order.
    document.body.offsetTop;
    document.getElementById("contentAddedAtTheBottom").classList.remove("hidden");
    document.body.offsetTop;
    document.getElementById("contentRemoved").classList.add("hidden");
}

function test()
{
    InspectorTest.initializeInspectorModels();

    var contentFlow;

    WebInspector.frameResourceManager.addEventListener(WebInspector.FrameResourceManager.Event.MainFrameDidChange, function(event) {
        var domTree = WebInspector.frameResourceManager.mainFrame.domTree;
        domTree.addEventListener(WebInspector.DOMTree.Event.RootDOMNodeInvalidated, onRootDOMNodeInvalidated, null);
        domTree.addEventListener(WebInspector.DOMTree.Event.ContentFlowWasAdded, onContentFlowWasAdded, null);
        domTree.requestContentFlowList();
    });

    function onRootDOMNodeInvalidated()
    {
        WebInspector.frameResourceManager.mainFrame.domTree.requestContentFlowList();
    }

    function onContentFlowWasAdded(event)
    {
        contentFlow = event.data.flow;
        InspectorTest.assert(contentFlow.name === "flow", "ContentFlow was added.");
        InspectorTest.assert(contentFlow.contentNodes.length == 2, "ContentFlow.contentNodes has a length of 2.");
        InspectorTest.assert(contentFlow.contentNodes[0].getAttribute("id") === "contentStatic", "ContentFlow.contentNodes[0].id is \"#contentStatic\".");
        InspectorTest.assert(contentFlow.contentNodes[1].getAttribute("id") === "contentRemoved", "ContentFlow.contentNodes[1].id is \"#contentRemoved\".");

        contentFlow.addEventListener(WebInspector.ContentFlow.Event.ContentNodeWasAdded, onContentNodeWasAdded, null);
        contentFlow.addEventListener(WebInspector.ContentFlow.Event.ContentNodeWasRemoved, onContentNodeWasRemoved, null);

        InspectorTest.sendCommand("Runtime.evaluate", {expression: "changeFlowContent()"});
    }

    function onContentNodeWasAdded(event)
    {
        switch (event.data.node.getAttribute("id")) {
        case "contentAddedAtTheTop":
            InspectorTest.assert(event.data.before.getAttribute("id") === "contentStatic", "\"#contentAddedAtTheTop\" was added before \"#contentStatic\".");
            InspectorTest.assert(contentFlow.contentNodes[0] === event.data.node, "\"#contentAddedAtTheTop\" is first in the contentNodes list.");
            break;
        case "contentAddedAtTheBottom":
            InspectorTest.assert(!event.data.before, "\"#contentAddedAtTheBottom\" was added last.");
            InspectorTest.assert(contentFlow.contentNodes.lastValue === event.data.node, "\"#contentAddedAtTheBottom\" is last in the contentNodes list.");
            break;
        default:
            InspectorTest.log("FAIL: Only two add events are expected.");
        }
        
    }

    function onContentNodeWasRemoved(event)
    {
        InspectorTest.assert(event.data.node.getAttribute("id") === "contentRemoved", "\"#contentRemoved\" was removed.");
        InspectorTest.assert(contentFlow.contentNodes.indexOf(event.data.node) === -1, "\"#contentRemoved\" cannot be found in the contentNodes list.");
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
    <p>Testing that the ContentFlows events are correctly dispatched when nodes are attached to/detached from a flow.</p>
    
    <div id="contentAddedAtTheTop" class="content hidden"></div>
    <div id="contentStatic" class="content"></div>
    <div id="contentRemoved" class="content"></div>
    <div id="contentAddedAtTheBottom" class="content hidden"></div>

</body>
</html>
